# fix(performance): Analyse and fix performance deopt in node.js