## Before
| library                     | relative speed | operations per second | avg. operation time |
| --------------------------- | -------------: | --------------------: | ------------------: |
| **@typeofweb/schema@0.4.2** |         **0%** |   **(1,187,900 rps)** |  **(avg: 0.841μs)** |
| mschema@0.5.6               |        -51.96% |         (570,698 rps) |          (avg: 1μs) |
| validator.js@2.0.4          |        -58.61% |         (491,630 rps) |          (avg: 2μs) |
| validate.js@0.13.1          |        -79.14% |         (247,790 rps) |          (avg: 4μs) |
| validatorjs@3.22.1          |        -85.98% |         (166,501 rps) |          (avg: 6μs) |
| joi@17.3.0                  |        -89.58% |         (123,763 rps) |          (avg: 8μs) |
| superstruct@0.13.3          |        -94.69% |          (63,099 rps) |         (avg: 15μs) |
| yup@0.32.8                  |        -94.64% |          (63,711 rps) |         (avg: 15μs) |
| jsonschema@1.4.0            |        -96.85% |          (37,471 rps) |         (avg: 26μs) |
| parambulator@1.5.2          |        -98.08% |          (22,823 rps) |         (avg: 43μs) |

## After
| library                     | relative speed | operations per second | avg. operation time |
| --------------------------- | -------------: | --------------------: | ------------------: |
| **@typeofweb/schema@0.4.2** |         **0%** |   **(1,443,686 rps)** |  **(avg: 0.692μs)** |
| mschema@0.5.6               |        -58.95% |         (592,589 rps) |          (avg: 1μs) |
| validator.js@2.0.4          |        -66.32% |         (486,197 rps) |          (avg: 2μs) |
| validate.js@0.13.1          |        -84.03% |         (230,496 rps) |          (avg: 4μs) |
| validatorjs@3.22.1          |        -88.67% |         (163,587 rps) |          (avg: 6μs) |
| joi@17.3.0                  |        -91.15% |         (127,809 rps) |          (avg: 7μs) |
| superstruct@0.13.3          |        -95.59% |          (63,604 rps) |         (avg: 15μs) |
| yup@0.32.8                  |        -96.06% |          (56,823 rps) |         (avg: 17μs) |
| jsonschema@1.4.0            |        -97.24% |          (39,834 rps) |         (avg: 25μs) |
| parambulator@1.5.2          |        -98.38% |          (23,428 rps) |         (avg: 42μs) |